+++ multiqc/modules/bcl2fastq/bcl2fastq.py
@@ -474,7 +474,7 @@ class MultiqcModule(BaseMultiqcModule):
             "id": "bcl2fastq-lane-stats-table",
             "table_title": "bcl2fastq Lane Statistics",
             "col1_header": "Run ID - Lane",
-            "no_beeswarm": True,
+            "no_violin": True,
         }
         return table.plot(self.bcl2fastq_bylane, headers, table_config)